+++ src/test/php/ImageTestCase.php
@@ -74,9 +74,9 @@ class ImageTestCase extends \PHPUnit_Framework_TestCase
     {
         $this->handle = imagecreatefrompng($this->testPath . 'empty.png');
         $image  = new Image('foo', null, $this->handle);
-        $this->assertEquals('foo', $image->getName());
-        $this->assertSame(ImageType::$PNG, $image->getType());
-        $this->assertSame($this->handle, $image->getHandle());
+        $this->assertEquals('foo', $image->fileName());
+        $this->assertSame(ImageType::$PNG, $image->type());
+        $this->assertSame($this->handle, $image->handle());
         $this->assertEquals('.png', $image->fileExtension());
         $this->assertEquals('image/png', $image->mimeType());
     }
@@ -87,9 +87,9 @@ class ImageTestCase extends \PHPUnit_Framework_TestCase
     public function instantiateWithLoad()
     {
         $image = Image::load($this->testPath . 'empty.png');
-        $this->assertEquals($this->testPath . 'empty.png', $image->getName());
-        $this->assertSame(ImageType::$PNG, $image->getType());
-        $this->assertNotNull($image->getHandle());
+        $this->assertEquals($this->testPath . 'empty.png', $image->fileName());
+        $this->assertSame(ImageType::$PNG, $image->type());
+        $this->assertNotNull($image->handle());
         $this->assertEquals('.png', $image->fileExtension());
         $this->assertEquals('image/png', $image->mimeType());
     }
@@ -101,9 +101,9 @@ class ImageTestCase extends \PHPUnit_Framework_TestCase
     public function instantiateWithLoadFromResource()
     {
         $image = Image::loadFromResource($this->testPath . 'empty.png', new ResourceLoader());
-        $this->assertEquals($this->testPath . 'empty.png', $image->getName());
-        $this->assertSame(ImageType::$PNG, $image->getType());
-        $this->assertNotNull($image->getHandle());
+        $this->assertEquals($this->testPath . 'empty.png', $image->fileName());
+        $this->assertSame(ImageType::$PNG, $image->type());
+        $this->assertNotNull($image->handle());
         $this->assertEquals('.png', $image->fileExtension());
         $this->assertEquals('image/png', $image->mimeType());
     }
@@ -117,8 +117,6 @@ class ImageTestCase extends \PHPUnit_Framework_TestCase
         $this->assertSame($image, $image->store('bar'));
         $this->assertEquals('bar', ImageType::$DUMMY->value()->lastStoredFileName());
         $this->assertSame($this->handle, ImageType::$DUMMY->value()->lastStoredHandle());
-        $this->assertEquals('.dummy', $image->fileExtension());
-        $this->assertEquals('image/dummy', $image->mimeType());
     }
 
     /**
@@ -129,7 +127,5 @@ class ImageTestCase extends \PHPUnit_Framework_TestCase
         $image = new Image('foo', ImageType::$DUMMY, $this->handle);
         $image->display();
         $this->assertSame($this->handle, ImageType::$DUMMY->value()->lastDisplayedHandle());
-        $this->assertEquals('.dummy', $image->fileExtension());
-        $this->assertEquals('image/dummy', $image->mimeType());
     }
 }